WIP
Onboarding page is seen once when the app is intalled.

Most states are working correcting now.

For example if you once you click get started you dont get shown the onboarding again.

When the app is loaded it will check if the user is signed in or not and show home